+++ HospitalityPro/Controllers/UserController.cs
@@ -2,7 +2,7 @@
 using DTO.UserDTO;
 using Microsoft.AspNetCore.Mvc;
 
-namespace HumanResourceProject.Controllers
+namespace HospitalityPro.Controllers
 
 {
     [ApiController]
