+++ HospitalityPro/appsettings.json
